requirements.txt
fastapi
uvicorn
transformers
torch
faiss-cpu
Pillow
python-multipart